+++ src/main/java/amp/awec/command/operation/CommandCopy.java
@@ -1,6 +1,6 @@
 package amp.awec.command.operation;
 
-import amp.awec.command.CommandHelper;
+import amp.awec.command.CommandPlayerData;
 import amp.awec.util.Vec3i;
 import amp.awec.data.PlayerData;
 import amp.awec.permission.WorldEditPermissions;
@@ -22,15 +22,16 @@ public class CommandCopy implements CommandManager.CommandRegistry {
 				.requires(source -> WorldEditPermissions.canUseWorldEdit((CommandSource) source))
 				.executes(context -> {
 					CommandSource source = (CommandSource) context.getSource();
-					PlayerData playerData = CommandHelper.getPlayerData(source);
+					CommandPlayerData playerData = CommandPlayerData.get(source);
 					if (playerData == null) {
 						return 0;
 					}
 
-					World world = source.getWorld();
-					Vec3i copyPos = PosHelper.getPlayerBlockPos(playerData.parentPlayer);
-					doCopy(world, playerData.selection, copyPos, playerData);
+					CuboidVolume selection = playerData.getSelection();
+					Vec3i copyPos = PosHelper.getPlayerBlockPos(playerData.player);
+					doCopy(playerData.world, selection, copyPos, playerData.data);
 					source.sendMessage("Copied");
+
 					return 1;
 				})
 		);