+++ src/main/java/amp/awec/command/operation/CommandWalls.java
@@ -1,9 +1,8 @@
 package amp.awec.command.operation;
-import amp.awec.command.CommandHelper;
+import amp.awec.command.CommandPlayerData;
 import amp.awec.operation.WorldChange;
 import amp.awec.operation.WallsOperation;
 import amp.awec.command.argtypes.ArgumentTypePattern;
-import amp.awec.data.PlayerData;
 import amp.awec.permission.WorldEditPermissions;
 import amp.awec.pattern.BlockPattern;
 import com.mojang.brigadier.CommandDispatcher;
@@ -13,7 +12,6 @@ import com.mojang.brigadier.builder.ArgumentBuilderRequired;
 import com.mojang.brigadier.context.CommandContext;
 import net.minecraft.core.net.command.CommandManager;
 import net.minecraft.core.net.command.CommandSource;
-import net.minecraft.core.world.World;
 
 public class CommandWalls implements CommandManager.CommandRegistry {
 
@@ -38,16 +36,15 @@ public class CommandWalls implements CommandManager.CommandRegistry {
 
 	private int handleWallsCommand(CommandContext<Object> context, int thickness) {
 		CommandSource source = (CommandSource) context.getSource();
-		PlayerData playerData = CommandHelper.getPlayerData(source);
+		CommandPlayerData playerData = CommandPlayerData.get(source);
 		if (playerData == null) {
 			return 0;
 		}
 
 		BlockPattern pattern = context.getArgument("pattern", BlockPattern.class);
 
-		World world = source.getWorld();
-		WorldChange result = WallsOperation.execute(world, playerData.selection, pattern, thickness);
-		playerData.undoHistory.add(result);
+		WorldChange result = WallsOperation.execute(playerData.world, playerData.getSelection(), pattern, thickness);
+		playerData.addUndoChange(result);
 
 		source.sendMessage("Changed " + result.changedBlockCount + " blocks");