+++ src/main/java/amp/awec/command/selection/CommandPos2.java
@@ -7,6 +7,6 @@ import net.minecraft.core.net.command.CommandSource;
 public class CommandPos2 implements CommandManager.CommandRegistry {
 	@Override
 	public void register(CommandDispatcher<CommandSource> dispatcher) {
-		CommandPosBase.register(dispatcher, "/pos2", 2, (data, pos) -> data.selection.setCorner2(pos));
+		CommandPosBase.register(dispatcher, "/pos2", 2, (data, pos) -> data.getSelection().setCorner2(pos));
 	}
 }