+++ src/main/java/amp/awec/command/selection/CommandPosBase.java
@@ -1,5 +1,7 @@
 package amp.awec.command.selection;
 
+import amp.awec.command.CommandPlayerData;
+import amp.awec.data.PlayerDataManager;
 import amp.awec.util.Vec3i;
 import amp.awec.WorldEditMod;
 import amp.awec.data.PlayerData;
@@ -18,38 +20,41 @@ import java.util.function.BiConsumer;
 public class CommandPosBase {
 
 	@SuppressWarnings("unchecked")
-	public static void register(CommandDispatcher<CommandSource> dispatcher, String command, int cornerNumber, BiConsumer<PlayerData, Vec3i> setter) {
+	public static void register(CommandDispatcher<CommandSource> dispatcher, String command, int cornerNumber, BiConsumer<CommandPlayerData, Vec3i> setter) {
 		dispatcher.register(
 			(ArgumentBuilderLiteral) ArgumentBuilderLiteral.literal(command)
 				.requires(source -> WorldEditPermissions.canUseWorldEdit((CommandSource) source))
 				.executes(context -> {
 					CommandSource source = (CommandSource) context.getSource();
-					Player player = source.getSender();
-					if (player != null) {
-						PlayerData playerData = WorldEditMod.getPlayerData(player);
-						Vec3i pos = PosHelper.getPlayerBlockPos(player);
-						setter.accept(playerData, pos);
-						source.sendMessage("Corner " + cornerNumber + " set to " + pos);
-						return 1;
+					CommandPlayerData playerData = CommandPlayerData.get(source);
+					if (playerData == null) {
+						return 0;
 					}
-					return 0;
+
+					Vec3i pos = PosHelper.getPlayerBlockPos(playerData.player);
+					setter.accept(playerData, pos);
+					source.sendMessage("Corner " + cornerNumber + " set to " + pos);
+
+					return 1;
 				})
 				.then(ArgumentBuilderRequired.argument("position", ArgumentTypeIntegerCoordinates.intCoordinates())
 					.executes(context -> {
 						CommandSource source = (CommandSource) context.getSource();
-						Player player = source.getSender();
-						if (player != null) {
-							IntegerCoordinates coordinates = context.getArgument("position", IntegerCoordinates.class);
-							int x = coordinates.getX(source);
-							int y = coordinates.getY(source, true);
-							int z = coordinates.getZ(source);
-							PlayerData playerData = WorldEditMod.getPlayerData(player);
-							Vec3i pos = new Vec3i(x, y, z);
-							setter.accept(playerData, pos);
-							source.sendMessage("Corner " + cornerNumber + " set to " + pos);
-							return 1;
+						CommandPlayerData playerData = CommandPlayerData.get(source);
+						if (playerData == null) {
+							return 0;
 						}
-						return 0;
+
+						IntegerCoordinates coordinates = context.getArgument("position", IntegerCoordinates.class);
+						int x = coordinates.getX(source);
+						int y = coordinates.getY(source, true);
+						int z = coordinates.getZ(source);
+
+						Vec3i pos = new Vec3i(x, y, z);
+						setter.accept(playerData, pos);
+						source.sendMessage("Corner " + cornerNumber + " set to " + pos);
+
+						return 1;
 				}))
 		);
 	}