+++ src/main/java/amp/awec/mixin/WandPos1ClientMixin.java
@@ -1,5 +1,6 @@
 package amp.awec.mixin;
 
+import amp.awec.data.PlayerDataManager;
 import amp.awec.util.Vec3i;
 import amp.awec.WorldEditMod;
 import amp.awec.data.PlayerData;
@@ -29,17 +30,15 @@ public class WandPos1ClientMixin {
 	private void startDestroyBlock(int x, int y, int z, Side side, double xHit, double yHit, boolean repeat, CallbackInfo ci) {
 		Player player = mc.thePlayer;
 
-		if (!WorldEditPermissions.canUseWorldEdit(player) || !WandHelper.isHoldingWand(player)) {
+		if (!WorldEditPermissions.canUseWorldEdit(player) || !WandHelper.isHoldingWand(player) || player.world == null) {
 			return;
 		}
 
-		PlayerData playerData = WorldEditMod.getPlayerData(player);
-		if (playerData == null) {
-			return;
-		}
+		PlayerData playerData = PlayerDataManager.getPlayerData(player.uuid);
 
 		Vec3i pos = new Vec3i(x, y, z);
-		playerData.selection.setCorner1(pos);
+		playerData.getSelection(player.world).setCorner1(pos);
+
 		player.sendMessage("Corner 1" + " set to " + pos);
 		ci.cancel();
 	}