+++ src/main/java/amp/awec/mixin/WandPos1ServerMixin.java
@@ -1,5 +1,6 @@
 package amp.awec.mixin;
 
+import amp.awec.data.PlayerDataManager;
 import amp.awec.util.Vec3i;
 import amp.awec.WorldEditMod;
 import amp.awec.data.PlayerData;
@@ -25,17 +26,14 @@ public class WandPos1ServerMixin {
 
 	@Inject(method = "startMining", at = @At("HEAD"), cancellable = true)
 	private void startMining(int x, int y, int z, Side side, CallbackInfo ci) {
-		if (!WorldEditPermissions.canUseWorldEdit(player) || !WandHelper.isHoldingWand(player)) {
+		if (!WorldEditPermissions.canUseWorldEdit(player) || !WandHelper.isHoldingWand(player) || player.world == null) {
 			return;
 		}
 
-		PlayerData playerData = WorldEditMod.getPlayerData(player);
-		if (playerData == null) {
-			return;
-		}
+		PlayerData playerData = PlayerDataManager.getPlayerData(player.uuid);
 
 		Vec3i pos = new Vec3i(x, y, z);
-		playerData.selection.setCorner1(pos);
+		playerData.getSelection(player.world).setCorner1(pos);
 		player.sendMessage("Corner 1 set to " + pos);
 		ci.cancel();