+++ src/main/java/amp/awec/util/PosHelper.java
@@ -2,6 +2,7 @@ package amp.awec.util;
 
 import net.minecraft.core.entity.player.Player;
 import net.minecraft.core.util.phys.Vec3;
+import org.jetbrains.annotations.Nullable;
 
 public class PosHelper {
 	public static Vec3i getPlayerBlockPos(Player player) {